# Use atom slice