… contact_map/dask_runner.py
… contact_map/tests/test_contact_map.py
… examples/contact_map.ipynb
… examples/contact_map_without_atom_slice.ipynb