This uses atom_slice to cut down a trajectory before computing the neighborlist. This should speed up contactmap/frequency calculations for real systems
The code is not yet completely implemented, but is only tested at one point, and not yet benchmarked for now.

TODO:
 - [x] write documentation
 - [x] write proper tests
 - [x] update example
 - [x] do benchmarking

EDIT: found the first bug already (only converting the list/ not the numpy arrays)