@dwhswenson 2018-04-25
The Travis error is a py-3.4 error that has existed for a few weeks now. It looks like something goes wrong in the pandas install. That's not on you (exists independently of this PR), but you're more than welcome to investigate! I haven't had time.

On the CodeClimate stuff, I'm somewhat ambivalent. Things that better fit the current CodeClimate rules (relaxed from the defaults) are better, but I'm willing to make exceptions. Maybe try making `use_atom_slice` a class variable that defaults to `True`? That way we can turn it off for more thorough testing,  but you don't increase argument count. That also removes cognitive complexity from `ContactObject.__init__`. I'll take a closer look at the complaints about `ContactObject.contact_map` and offer suggestions later (possibly offline). Obviously, you intend to fix your `TODO`s, so I'm not worried about those.

Finally, don't worry about serializing `use_atom_slice`. I take the view that things that only affect the performance, not the actual output, don't need to be serialized. We expect that this should be that sort of situation.