@sroet 2018-05-15
Alright, I benchmarked everything in `contact_map.ipynb`. The atom slicing adds some overhead but is faster for most situations, on my system (default slices 2722 atoms to 1408):

- The single `atom_slice` per `ContactMap` or `ContactFrequency`: adds about 0.4 s. This scales with the number of atoms left after the slicing. It however does not scale with the length of the trajectory (as long as both the original and sliced trajectory fit in memory)

- The computation time of the neighborlist reduces by the relative amount of atoms sliced, from 0.02 s to 0.01 s per frame for the defaults

- The conversion of the neighborlist to the original indices adds about 0.01 s per frame for the default settings, this also time reduces if less atoms are left after slicing

Even for this trajectory without waters using atom_slice performs better for almost everything in the example, except for the `ContactMap` (110 ms to 516 ms because of the single 0.4s overhead) and the `ContactFrequency` with only a different query and no different haystack (2.87s to 3.18s). On the other hand for the example where both the query and haystack are defined (slicing the system down to 90 atoms) it is much faster (2.71s to 348ms). 

Seeing this I would argue that `use_atom_slice=True` should be the default for both `ContactMap` and `ContactFrequency`. What is your opinion on this @dwhswenson ? And could you see if these benchmarks hold up you?